requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==2.0.12
Django==3.2.16
django-bootstrap5==22.2
djangorestframework==3.12.4
idna==3.8
pytz==2024.2
requests==2.26.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==1.26.20
